harden: address code-review findings on the hardening commit
- archive: split the extraction budget into extracted (1 GiB, written) and
  inflated (4 GiB, scanned) ceilings so a legitimate archive with a small
  extracted output but large unused members still installs, while the
  decompression-bomb guard on total inflation stays; validate tar header
  sizes against the inflated ceiling.
- archive: a member covered by BOTH a file and a directory mapping now
  marks the directory prefix matched and is written to every mapped target
  (previously it false-failed with "archive directory not found").
- archive: raise the extraction time backstop 60s -> 300s so a large
  legitimate archive on a slow machine is not aborted; drop the dead
  isDirectory branch.
- cli: validate --parallelism in call() instead of via an @Spec field,
  which native-image never registers for field reflection (it NPE'd in the
  shipped binary, reintroducing the raw stack trace the fix prevented).

169 tests pass; plan/versions smoke tests verified.

+++ core/test/src/binstaller/core/CoreTestSupport.scala
@@ -467,7 +467,7 @@ private[core] trait CoreTestSupport:
     gzip.close()
     output.toByteArray
 
-  protected def tarHeader(name: String, size: Int, entryType: Char): Array[Byte] =
+  protected def tarHeader(name: String, size: Long, entryType: Char): Array[Byte] =
     val header = Array.fill[Byte](512)(0)
     writeTarField(header, 0, 100, name)
     writeTarField(header, 124, 12, f"$size%011o")
